// src/core/interfaces/orders-list-props.d.ts
import { Prisma } from "@prisma/client";

export interface IOrdersListProps {
  orders: Array<
    Prisma.OrderGetPayload<{
      include: {
        restaurant: {
          select: {
            name: true;
            avatarImageUrl: true;
          };
        };
        orderProducts: {
          include: {
            product: true;
          };
        };
      };
    }>
  >;
}